fuel-crisis-app/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
databricks-sql-connector>=3.0.0
databricks-sdk>=0.20.0
asyncpg>=0.29.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
